backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Pedro Santos <pedrosantoscrlx@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.34.0"
psycopg = { extras = ["binary"], version = "^3.2" }
sqlalchemy = "^2.0"
python-dotenv = "^1.1.0"
python-jose = "^3.5.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
bcrypt = "^3.2.2"
email-validator = "^2.2.0"
argon2-cffi = "^25.1.0"
requests = "^2.32.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
